requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==20.1.0
click==7.1.2
coverage==5.3
Flask==1.1.2
Flask-Cors==3.0.9
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
mistune==0.8.4
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
Werkzeug==1.0.1
